fix: narrow pooled embeddings after normalization, not inside pooling
The previous commit narrowed inside mean_pooling, which also moved the accumulation
and the subsequent normalize() into the input dtype. For float16 graph output that
overflows: a fully attended 1024-d vector of 10.0 squares to 102400, past float16's
65504 maximum, so the norm becomes inf and every returned component is exactly 0.

Keep mean_pooling's float64 accumulation exactly as upstream has it, and narrow at the
post-processing boundary instead - after normalize() for PooledNormalizedEmbedding.
The returned embedding now carries the dtype the ONNX graph produced, as every
non-pooled path already does, with no change to the arithmetic that produces it.

Tests cover float32, float16 and float64 through both pooled families, plus the
float16 overflow case directly. M14 nano parity is unchanged and slightly tighter:
minimum true cosine 1.0 across Torch/direct-ORT/FastEmbed with maximum absolute
error 1.1548399925231934e-07.

+++ fastembed/common/utils.py
@@ -29,7 +29,7 @@ def mean_pooling(input_array: NumpyArray, attention_mask: NDArray[np.int64]) -> 
     sum_embeddings = np.sum(input_array * input_mask_expanded, axis=1)
     sum_mask = np.sum(input_mask_expanded, axis=1)
     pooled_embeddings = sum_embeddings / np.maximum(sum_mask, 1e-9)
-    return pooled_embeddings.astype(input_array.dtype, copy=False)
+    return pooled_embeddings
 
 
 def last_token_pooling(input_array: NumpyArray, attention_mask: NDArray[np.int64]) -> NumpyArray:
